Add separate component for search location
